% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readadmb.R
\name{colr}
\alias{colr}
\title{Transparent colors}
\usage{
colr(col.pal = 1, a = 1)
}
\arguments{
\item{col.pal}{Color palette or vector of colors.}

\item{a}{Numeric alpha in \code{[0, 1]}.}
}
\value{
Character vector of colors with alpha applied.
}
\description{
Transparent colors
}
\note{
Author: Nathan Stephens (hacks package)
}
