% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{fn_plot_anoms}
\alias{fn_plot_anoms}
\title{Plot weight-at-age anomalies}
\usage{
fn_plot_anoms(dfin, maxage = 10, firstyr = 1982, minage = 3, txtsize = 3)
}
\arguments{
\item{dfin}{Data frame output from \code{fn_get_pred()}.}

\item{maxage}{Maximum age to plot.}

\item{firstyr}{First year to include.}

\item{minage}{Minimum age to plot.}

\item{txtsize}{Text label size.}
}
\value{
A ggplot object.
}
\description{
Plot weight-at-age anomalies
}
