% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readadmb.R
\name{read_dat}
\alias{read_dat}
\title{Read ADMB data files}
\usage{
read_dat(datfile)
}
\arguments{
\item{repfile}{ADMB output files to be read (no extension needed)}
}
\value{
object of type 'list' with ADMB outputs as list elements
}
\description{
Read ADMB data files
}
\author{
Steve Martell, Anders Nielsen, Athol Whitten, D'Arcy N. Webber
}
