% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{replace_zeros_with_mean}
\alias{replace_zeros_with_mean}
\title{Replace outliers with mean for mean weight-at-age}
\usage{
replace_zeros_with_mean(column)
}
\arguments{
\item{column}{Numeric vector.}
}
\value{
Numeric vector with outliers replaced by the mean of non-zero values.
}
\description{
Replace outliers with mean for mean weight-at-age
}
