% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{replace_zeros_with_two}
\alias{replace_zeros_with_two}
\title{Replace zeros with near-two for SD}
\usage{
replace_zeros_with_two(column)
}
\arguments{
\item{column}{Numeric vector.}
}
\value{
Numeric vector with zeros replaced by 1.9999.
}
\description{
Replace zeros with near-two for SD
}
