% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readadmb.R
\name{write_dat}
\alias{write_dat}
\title{Write ADMB data file}
\usage{
write_dat(A, datfile)
}
\arguments{
\item{A}{List of ADMB objects.}

\item{datfile}{Output data file path.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Write ADMB data file
}
