% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt.R
\docType{package}
\name{WtAgeRe-package}
\alias{WtAgeRe}
\alias{WtAgeRe-package}
\title{WtAgeRe}
\description{
Utilities for weight-at-age modeling and ADMB output processing.
}
\author{
\strong{Maintainer}: Jim Ianelli \email{Jim.Ianelli@noaa.gov}

}
\keyword{internal}
